Added user info api and showing data in profile screen

+++ app/src/main/java/de/westnordost/streetcomplete/data/workspace/WorkspaceDao.kt
@@ -33,7 +33,7 @@ class WorkspaceDao(private val db: Database) {
                 it.getString(TITLE))
         }
 
-    fun deleteAll(ids: List<Long>): Int {
+    fun deleteAll(ids: List<Int>): Int {
         if (ids.isEmpty()) return 0
         return db.delete(NAME, "$ID in (${ids.joinToString(",")})")
     }
